hass-backup/server/app/requirements.txt
click==7.1.1
Flask==1.1.2
Flask-HTTPAuth==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
pytz==2019.3
PyYAML==5.3.1
tzlocal==2.0.0
Werkzeug==1.0.1
